Divided the Pages to different controllers.
