+++ WebProject.Tests/UnitTest1.cs
@@ -1,0 +1,12 @@
+namespace WebProject.Tests
+{
+    [TestClass]
+    public class UnitTest1
+    {
+        // Moq FrameWork -- Optional
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
