+++ WebProject.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.35027.167
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebProject", "WebProject\WebProject.csproj", "{5F151945-777F-4CEB-8112-73586AF424FD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebProject", "WebProject\WebProject.csproj", "{5F151945-777F-4CEB-8112-73586AF424FD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebProject.Tests", "WebProject.Tests\WebProject.Tests.csproj", "{014D3C21-0EC0-4322-AE9E-62FB7D46D54D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{5F151945-777F-4CEB-8112-73586AF424FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5F151945-777F-4CEB-8112-73586AF424FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5F151945-777F-4CEB-8112-73586AF424FD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{014D3C21-0EC0-4322-AE9E-62FB7D46D54D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{014D3C21-0EC0-4322-AE9E-62FB7D46D54D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{014D3C21-0EC0-4322-AE9E-62FB7D46D54D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{014D3C21-0EC0-4322-AE9E-62FB7D46D54D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
