+++ WebProject/Controllers/ErrorController.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace WebProject.Controllers
+{
+    public class ErrorController : Controller
+    {
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
