+++ WebProject/Controllers/HomeController.cs
@@ -18,27 +18,27 @@ namespace WebProject.Controllers
             return View(topTwoAnimals);
         }
 
-        [HttpGet]
-        public IActionResult Catalog(string category)
-        {
-            var categories = _repository.GetCategories().Select(c => c.Name);
-            ViewBag.Categories = new SelectList(categories);
+        //[HttpGet]
+        //public IActionResult Catalog(string category)
+        //{
+        //    var categories = _repository.GetCategories().Select(c => c.Name);
+        //    ViewBag.Categories = new SelectList(categories);
 
-            IEnumerable<Animal> animals;
-            if (category == null)
-                animals = _repository.GetAnimals();
-            else
-                animals= _repository.GetAnimals(category);
-            return View(animals);
-        }
-        [HttpPost]
-        public IActionResult RedirectToCatalog(string category)
-        {
-            return RedirectToAction("Catalog", new {category = category});
-        }
-        public IActionResult Administrator()
-        {
-            return View();
-        }
+        //    IEnumerable<Animal> animals;
+        //    if (category == null)
+        //        animals = _repository.GetAnimals();
+        //    else
+        //        animals= _repository.GetAnimals(category);
+        //    return View(animals);
+        //}
+        //[HttpPost]
+        //public IActionResult RedirectToCatalog(string category)
+        //{
+        //    return RedirectToAction("Catalog", new {category = category});
+        //}
+        //public IActionResult Administrator()
+        //{
+        //    return View();
+        //}
     }
 }
